requirements.txt
google-auth==2.43.0
google-auth-oauthlib==1.2.1
python-dotenv==1.2.1
requests==2.32.5
openpyxl==3.1.5
pandas==2.3.3